pyproject.toml
[tool.poetry]
name = "document q&a rag app with gemma and groq api"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-groq = "^0.2.0"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
langchain-community = "^0.3.3"
faiss-cpu = "^1.9.0"
langchain-google-genai = "^2.0.1"
pypdf = "^5.0.1"
langchain = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
